Lazy load the reader when opening a `File` (#4)
There are several cases where we open a `File`, but don't actually need
to read the contents (e.g. to check metadata).

`zip.into_entry()` reads from the zip file internally in order to check
signatures. This can be wasteful if reads for the underlying data source
are expensive (e.g. if a seek + read triggers a new HTTP request).

Therefore, this PR modifies `File` to lazily call `zip.into_entry()` if
we actually need to read from the file.